requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.9
mysql==0.0.3
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
